Avoid using infra nodes on OSD when selecting workers.
On OSD, infra nodes are also labeled as worker nodes by a technical
limitation. This change will make it so that the scale testing will only
select worker nodes that are not also labeled as infra nodes. This
should be transparent for regular scale testing as the infra nodes are
not labeled this way, but should hopefully maintain parity with the
intended functionality on OSD.